[refactor] MoveCondition 수정
