+++ src/main/java/racingcar/condition/MoveCondition.java
@@ -1,8 +1,8 @@
-package racingcar.domain;
+package racingcar.condition;
 
 /**
  * 점수 조건
  */
 public interface MoveCondition {
-    boolean isMovable(int num);
+    boolean isMovable();
 }
