--- src/main/java/racingcar/domain/RandomMoveCondition.java
@@ -1,10 +1,0 @@
-package racingcar.domain;
-
-public class RandomMoveCondition implements MoveCondition {
-    private static final int MINIMUM_OF_MOVE = 4;
-
-    @Override
-    public boolean isMovable(int num) {
-        return num >= MINIMUM_OF_MOVE;
-    }
-}
