+++ src/main/java/racingcar/domain/Winners.java
@@ -1,0 +1,2 @@
+package racingcar.domain;public class Winners {
+}
